fix: DescriptionRawItem
