+++ packages/lab/src/index.ts
@@ -7,6 +7,7 @@ export * from './chat/Chat';
 export * from './chat/ChatMessage';
 export * from './container/Container';
 export * from './description-item/DescriptionItem';
+export * from './description-item/DescriptionRawItem';
 export * from './file-drop-zone/FileDropZone';
 export * from './file-list-item/FileListItem';
 export * from './linked-text/LinkedText';
