+++ packages/ui/src/description-list/DescriptionList.spec.tsx
@@ -21,9 +21,6 @@ test('label id', () => {
         title="Text"
       >
         <span
-          class="DescriptionList__PriceText-SD__sc-117p3cv-0 hwLMrW"
-        />
-        <span
           class="MuiTypography-root MuiTypography-body2 MuiTypography-colorTextSecondary"
           id="label-id"
         >
@@ -51,9 +48,6 @@ test('dynamic label id', () => {
         class="MuiTypography-root SD-OverflowText-root MuiTypography-body2 MuiTypography-colorTextPrimary MuiTypography-noWrap"
         title="Text"
       >
-        <span
-          class="DescriptionList__PriceText-SD__sc-117p3cv-0 hwLMrW"
-        />
         <span
           class="MuiTypography-root MuiTypography-body2 MuiTypography-colorTextSecondary"
           id="uid_2"
