+++ packages/ui/src/description-list/DescriptionList.stories.tsx
@@ -1,5 +1,4 @@
 import { Link } from '@material-ui/core';
-import { CreditCard } from '@material-ui/icons';
 import CalendarTodayIcon from '@material-ui/icons/CalendarToday';
 import ChatIcon from '@material-ui/icons/Chat';
 import NotesIcon from '@material-ui/icons/Notes';
@@ -10,7 +9,6 @@ import { Meta } from '@storybook/react';
 import { PhoneLink, PhoneText } from '@superdispatch/phones';
 import { Box } from '@superdispatch/ui-lab';
 import { DescriptionList, DescriptionListItem } from './DescriptionList';
-import { DescriptionRawItem } from './DescriptionRawItem';
 
 export default {
   title: 'Data Display/DescriptionList',
@@ -146,23 +144,5 @@ export const nested = () => (
       content=""
       fallback="No pickup notes"
     />
-  </DescriptionList>
-);
-
-export const payment = () => (
-  <DescriptionList>
-    <DescriptionListItem
-      icon={<CreditCard />}
-      price="$1200"
-      label="COD"
-      fallback="$0.5/mi"
-    />
-  </DescriptionList>
-);
-
-export const descriptionRaw = () => (
-  <DescriptionList>
-    <DescriptionRawItem title="Payment" value="$1200" boldValue={true} />
-    <DescriptionRawItem title="Method" value="Comcheck" />
   </DescriptionList>
 );
