+++ packages/ui/src/description-list/DescriptionList.tsx
@@ -2,16 +2,11 @@ import { SvgIcon, Typography, TypographyProps } from '@material-ui/core';
 import { CSSProperties, makeStyles } from '@material-ui/styles';
 import clsx from 'clsx';
 import { forwardRef, ReactNode } from 'react';
-import styled from 'styled-components';
 import { OverflowText, OverflowTextProps } from '../overflow-text/OverflowText';
 import { Color } from '../theme/Color';
 import { SuperDispatchTheme } from '../theme/SuperDispatchTheme';
 import { isEmptyReactNode } from '../utils/isEmptyReactNode';
 import { useUID } from '../utils/useUID';
-
-const PriceText = styled.span`
-  color: ${Color.Purple500};
-`;
 
 function sizeVariant(
   theme: SuperDispatchTheme,
@@ -103,7 +98,6 @@ export interface DescriptionListItemProps {
   content?: ReactNode;
   contentTypographyProps?: Omit<OverflowTextProps, 'component' | 'color'>;
 
-  price?: ReactNode;
   fallback?: ReactNode;
 }
 
@@ -122,7 +116,6 @@ export const DescriptionListItem = forwardRef<
       content,
       contentTypographyProps,
 
-      price,
       fallback,
     },
     ref,
@@ -144,7 +137,7 @@ export const DescriptionListItem = forwardRef<
           {...contentTypographyProps}
           component="span"
           color={
-            (shouldRenderFallback && label == null) || price
+            shouldRenderFallback && label == null
               ? 'textSecondary'
               : 'textPrimary'
           }
@@ -156,15 +149,13 @@ export const DescriptionListItem = forwardRef<
               undefined,
           }}
         >
-          {price !== null && <PriceText>{price}</PriceText>}
-          {price != null && ' '}
           {label != null && (
             <Typography
               {...labelTypographyProps}
               id={labelID}
               variant="body2"
               component="span"
-              color={price == null ? 'textSecondary' : 'textPrimary'}
+              color="textSecondary"
             >
               {label}
             </Typography>
