+++ packages/ui/src/index.ts
@@ -8,7 +8,6 @@ export * from './checkbox/CheckboxGroudField';
 export * from './columns/Column';
 export * from './columns/Columns';
 export * from './description-list/DescriptionList';
-export * from './description-list/DescriptionRawItem';
 export * from './drawer/DrawerActions';
 export * from './drawer/DrawerContent';
 export * from './drawer/DrawerList';
